refactor: UserService 기능 분할
