+++ backend/src/main/java/com/chakak/filter/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
 package com.chakak.filter;
 
-import com.chakak.service.UserService;
+import com.chakak.service.AuthService;
 import com.chakak.util.JwtUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -22,7 +22,7 @@ import java.io.IOException;
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
 	private final JwtUtil jwtUtil;
-	private final UserService userService;
+	private final AuthService authService;
 
 	@Override
 	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
@@ -48,7 +48,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
 			if (username != null && SecurityContextHolder.getContext().getAuthentication() == null) {
 				try {
-					UserDetails userDetails = userService.loadUserByUsername(username);
+					UserDetails userDetails = authService.loadUserByUsername(username);
 
 					if (jwtUtil.validateToken(token, userDetails)) {
 						UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
